sweep of templates.
