+++ templates/babonus.hbs
@@ -44,7 +44,7 @@
         </div>
 
         <div class="utils">
-          <div class="notification warning" id="babonus-warning"></div>
+          <div class="notification warning" id="babonus-warning">{{{localize "BABONUS.WARNINGS.INVALID"}}}</div>
           <div class="utils-buttons">
             <button data-type="save-button" type="submit">
               <i class="fa-solid fa-floppy-disk"></i> {{localize "BABONUS.BUTTONS.SAVE"}}
