+++ templates/builder_components/_required_fields.hbs
@@ -10,7 +10,7 @@
     <span class="sep">{{localize "BABONUS.LABELS.NAME"}}</span>
     <input type="text" name="name" placeholder="{{localize 'BABONUS.PLACEHOLDERS.NAME'}}">
     <span class="sep">{{localize "BABONUS.LABELS.ID"}}</span>
-    <input type="text" name="id" placeholder="{{localize 'BABONUS.PLACEHOLDERS.ID'}}" disabled>
+    <input type="text" name="id" value="{{id}}" placeholder="{{localize 'BABONUS.PLACEHOLDERS.ID'}}" disabled>
   </div>
 </div>
 <div class="form-group">
