+++ templates/builder_components/text_keys.hbs
@@ -15,7 +15,7 @@
   </label>
   <div class="form-fields">
     <input type="text" name="filters.{{name}}">
-    <button class="babonus-keys" data-type="{{name}}" type="button">
+    <button class="babonus-keys" type="button">
       <i class="fa-solid fa-key"></i> {{localize "BABONUS.BUTTONS.KEYS"}}
     </button>
   </div>
