+++ templates/builder_components/text_text_keys.hbs
@@ -8,7 +8,7 @@
   <div class="form-fields">
     <input type="text" name="filters.{{name}}.needed">
     <input type="text" name="filters.{{name}}.unfit">
-    <button class="babonus-keys" data-type="{{name}}" type="button">
+    <button class="babonus-keys" type="button">
       <i class="fa-solid fa-key"></i> {{localize "BABONUS.BUTTONS.KEYS"}}
     </button>
   </div>
