--- templates/keys_spellComponents.hbs
@@ -1,19 +1,0 @@
-<p class="babonus-key-description">
-  {{localize "BABONUS.KEY.COMPONENTS_DESC"}}
-</p>
-<table class="babonus two-col">
-  <thead>
-    <tr>
-      <th>{{localize "BABONUS.KEY.COMPONENTS_NAME"}}</th>
-      <th><a class="select-all" data-index="2">{{localize "BABONUS.KEY.COMPONENTS_BOX"}}</a></th>
-    </tr>
-  </thead>
-  <tbody>
-    {{#each types as |type|}}
-    <tr>
-      <td>{{type.label}}</td>
-      <td><input type="checkbox" id="{{type.value}}" {{checked type.checked}}></td>
-    </tr>
-    {{/each}}
-  </tbody>
-</table>
