--- templates/keys_targetEffects.hbs
@@ -1,19 +1,0 @@
-<p class="babonus-key-description">
-  {{localize "BABONUS.KEY.TARGET_CONDITIONS_DESC"}}
-</p>
-<table class="babonus two-col has-image">
-  <thead>
-    <tr>
-      <th>{{localize "BABONUS.KEY.TARGET_CONDITIONS_NAME"}}</th>
-      <th><a class="select-all" data-index="2">{{localize "BABONUS.KEY.TARGET_CONDITIONS_BOX"}}</a></th>
-    </tr>
-  </thead>
-  <tbody>
-    {{#each types as |type|}}
-    <tr>
-      <td><img src="{{type.icon}}">{{type.label}}</td>
-      <td><input type="checkbox" id="{{type.value}}" {{checked type.checked}}></td>
-    </tr>
-    {{/each}}
-  </tbody>
-</table>
