--- templates/keys_weaponProperties.hbs
@@ -1,21 +1,0 @@
-<p class="babonus-key-description">
-  {{localize "BABONUS.KEY.PROPERTIES_DESC"}}
-</p>
-<table class="babonus three-col">
-  <thead>
-    <tr>
-      <th>{{localize "BABONUS.KEY.PROPERTIES_NAME"}}</th>
-      <th><a class="select-all" data-index="2">{{localize "BABONUS.KEY.PROPERTIES_NEED"}}</a></th>
-      <th><a class="select-all" data-index="3">{{localize "BABONUS.KEY.PROPERTIES_UNFIT"}}</a></th>
-    </tr>
-  </thead>
-  <tbody>
-    {{#each types as |type|}}
-    <tr>
-      <td>{{type.label}}</td>
-      <td><input type="checkbox" id="{{type.value}}" data-property="needed" {{checked type.needed}}></td>
-      <td><input type="checkbox" id="{{type.value}}" data-property="unfit" {{checked type.unfit}}></td>
-    </tr>
-    {{/each}}
-  </tbody>
-</table>
