+++ templates/subapplications/keysSingle.hbs
@@ -1,0 +1,19 @@
+<p class="babonus-key-description">
+  {{localize description}}
+</p>
+<table class="babonus two-col">
+  <thead>
+    <tr>
+      <th>{{localize head1}}</th>
+      <th><a class="select-all">{{localize head2}}</a></th>
+    </tr>
+  </thead>
+  <tbody>
+    {{#each types as |type|}}
+    <tr>
+      <td>{{#if type.icon}}<img src="{{type.icon}}">{{/if}}{{type.label}}</td>
+      <td><input type="checkbox" id="{{type.value}}" {{checked type.checked}}></td>
+    </tr>
+    {{/each}}
+  </tbody>
+</table>
